[B] Fix text collaborator sorting in updater
Fixes #929
