+++ api/app/lib/updaters/text.rb
@@ -2,6 +2,7 @@ module Updaters
   # Updates a Text model from JSON-API style params
   class Text
     include ::Updaters
+    include ::Updaters::Concerns::HasSortableCollaborators
 
     def update(model)
       @model = model
